+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.1.0 - 2024-02-23]
+## [1.1.0 - 2024-02-24]
 ### Added
 - Dockerfile
 - Single Spa
@@ -26,3 +26,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - If type is DataSHIELD, use template opal-ccp and output format Opal
 - Creator in main dashboard
 - Invite users to project
+- Downloads